Add index.js with useFccTest and FccTests
`useFccTest` is a React hook to display the FCC tests.
It can accept two optional parameters, `fccTest` and `queryParam`.
`fccTest` is used to specify which test should be selected.
`queryParam` is used to only display the FCC tests if the
specificed query parameter is present.

`FccTests` is an object which contains a list of the FCC tests.
A selected item can be passed to `useFccTest` with the `fccTest`
parameter.